refactor(layering): relocate ICRegistry from Element/ to Scene/
ICRegistry is a scene-level service (per-scene blob store + file watcher,
pushes undo commands, reloads IC instances, holds a Scene*). Living under
App/Element/ made it the source of several Element->{Scene,Commands,
Simulation} upward includes; moving it to App/Scene/ makes those dependencies
intra-layer / downward.

Pure relocation: git mv + include-path updates across 18 files and
CMakeSources, plus the TestDanglingPointer source-path string.

Behaviour-neutral; full suite 176/176 green.
