+++ App/Scene/SceneDropHandler.cpp
@@ -15,9 +15,9 @@
 #include "App/Element/ElementFactory.h"
 #include "App/Element/GraphicElement.h"
 #include "App/Element/IC.h"
-#include "App/Element/ICRegistry.h"
 #include "App/IO/Serialization.h"
 #include "App/Scene/Commands.h"
+#include "App/Scene/ICRegistry.h"
 #include "App/Scene/Scene.h"
 
 SceneDropHandler::SceneDropHandler(Scene *scene)
