+++ App/Scene/Workspace.cpp
@@ -35,7 +35,6 @@ bool isReadOnlyFailure(QFileDevice::FileError error)
 #include "App/Core/Settings.h"
 #include "App/Element/GraphicElement.h"
 #include "App/Element/IC.h"
-#include "App/Element/ICRegistry.h"
 #include "App/IO/FileUtils.h"
 #include "App/IO/Serialization.h"
 #include "App/IO/SerializationContext.h"
@@ -43,6 +42,7 @@ bool isReadOnlyFailure(QFileDevice::FileError error)
 #include "App/Nodes/QNEConnection.h"
 #include "App/Nodes/QNEPort.h"
 #include "App/Scene/Commands.h"
+#include "App/Scene/ICRegistry.h"
 #include "App/Simulation/SimulationBlocker.h"
 #include "App/UI/FileDialogProvider.h"
 #include "App/Versions.h"
