+++ App/UI/ElementEditor.cpp
@@ -20,10 +20,10 @@
 #include "App/Element/GraphicElements/Node.h"
 #include "App/Element/GraphicElements/TruthTable.h"
 #include "App/Element/IC.h"
-#include "App/Element/ICRegistry.h"
 #include "App/IO/Serialization.h"
 #include "App/Nodes/QNEConnection.h"
 #include "App/Scene/Commands.h"
+#include "App/Scene/ICRegistry.h"
 #include "App/Scene/Scene.h"
 #include "App/UI/ElementContextMenu.h"
 #include "App/UI/ElementEditorUI.h"
