+++ App/UI/ICController.cpp
@@ -16,9 +16,9 @@
 #include "App/Core/Enums.h"
 #include "App/Core/SentryHelpers.h"
 #include "App/Element/IC.h"
-#include "App/Element/ICRegistry.h"
 #include "App/IO/Serialization.h"
 #include "App/Scene/Commands.h"
+#include "App/Scene/ICRegistry.h"
 #include "App/Scene/Scene.h"
 #include "App/Scene/Workspace.h"
 #include "App/Simulation/Simulation.h"
