+++ App/UI/MainWindow.cpp
@@ -39,9 +39,9 @@
 #include "App/Element/ElementLabel.h"
 #include "App/Element/IC.h"
 #include "App/Element/ICPreviewPopup.h"
-#include "App/Element/ICRegistry.h"
 #include "App/IO/RecentFiles.h"
 #include "App/Scene/GraphicsView.h"
+#include "App/Scene/ICRegistry.h"
 #include "App/Scene/Workspace.h"
 #include "App/Simulation/Simulation.h"
 #include "App/UI/ElementPalette.h"
