+++ App/UI/SceneUiBinder.cpp
@@ -13,8 +13,8 @@
 #include "App/Core/Common.h"
 #include "App/Element/IC.h"
 #include "App/Element/ICPreviewPopup.h"
-#include "App/Element/ICRegistry.h"
 #include "App/Scene/GraphicsView.h"
+#include "App/Scene/ICRegistry.h"
 #include "App/Scene/Scene.h"
 #include "App/Scene/Workspace.h"
 #include "App/Simulation/Simulation.h"
