+++ Tests/Fuzz/FuzzRoundTrip.cpp
@@ -35,8 +35,8 @@
 #include "App/Core/Common.h"
 #include "App/Core/Enums.h"
 #include "App/Element/GraphicElement.h"
-#include "App/Element/ICRegistry.h"
 #include "App/IO/Serialization.h"
+#include "App/Scene/ICRegistry.h"
 #include "App/Scene/Scene.h"
 #include "App/Scene/Workspace.h"
 #include "App/Simulation/Simulation.h"
