+++ Tests/Integration/TestArduino.cpp
@@ -31,10 +31,10 @@ QString TestArduino::s_cliCachePath;
 #include "App/Element/GraphicElements/Led.h"
 #include "App/Element/GraphicElements/Node.h"
 #include "App/Element/IC.h"
-#include "App/Element/ICRegistry.h"
 #include "App/Core/Priorities.h"
 #include "App/Nodes/QNEConnection.h"
 #include "App/Nodes/QNEPort.h"
+#include "App/Scene/ICRegistry.h"
 #include "App/Scene/Workspace.h"
 #include "App/Simulation/Simulation.h"
 #include "Tests/Common/ICTestHelpers.h"
