+++ Tests/Integration/TestMainWindowGui.cpp
@@ -28,12 +28,12 @@
 #include "App/Element/GraphicElements/Node.h"
 #include "App/Element/GraphicElements/Or.h"
 #include "App/Element/IC.h"
-#include "App/Element/ICRegistry.h"
 #include "App/IO/Serialization.h"
 #include "App/Nodes/QNEConnection.h"
 #include "App/Nodes/QNEPort.h"
 #include "App/Scene/Commands.h"
 #include "App/Scene/GraphicsView.h"
+#include "App/Scene/ICRegistry.h"
 #include "App/Scene/Scene.h"
 #include "App/Scene/Workspace.h"
 #include "App/Simulation/Simulation.h"
